refactor: UserService 인터페이스화
